% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acutelines.datatoolbox-package.R
\docType{package}
\name{acutelines.datatoolbox-package}
\alias{acutelines.datatoolbox}
\alias{acutelines.datatoolbox-package}
\title{acutelines.datatoolbox: Toolbox to work with Acutelines biobank data.}
\description{
This package can be used to work with data extracted from the Acutelines data- and biobank. It provides tools to parse data extracts and perform calculations in a uniform matter.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://umcg-seh.github.io/acutelines.datatoolbox/}
  \item \url{https://github.com/UMCG-SEH/acutelines.datatoolbox}
  \item Report bugs at \url{https://github.com/UMCG-SEH/acutelines.datatoolbox/issues}
}

}
\author{
\strong{Maintainer}: Raymond J. van Wijk \email{r.j.van.wijk@umcg.nl} (\href{https://orcid.org/0000-0002-2106-7699}{ORCID}) (Department of Acute Care, University of Groningen, University Medical Center Groningen, Groningen, The Netherlands)

Authors:
\itemize{
  \item Carolina Hincapié-Osorno (\href{https://orcid.org/0000-0001-8582-1614}{ORCID}) (Department of Internal Medicine, University of Groningen, University Medical Center Groningen, Groningen, The Netherlands)
}

Other contributors:
\itemize{
  \item Tamar van der Aart [contributor]
  \item Sanne Ter Horst [contributor]
  \item Acutelines, UMCG \email{acutelines@umcg.nl} [copyright holder]
}

}
\keyword{internal}
