% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{cbrt}
\alias{cbrt}
\title{Calculate cube root of x}
\usage{
cbrt(x)
}
\arguments{
\item{x}{x}
}
\value{
x^(1/3)
}
\description{
Calculate cube root of x
}
\seealso{
Other helper: 
\code{\link{impute_fio2}()},
\code{\link{impute_pao2}()},
\code{\link{o2supply_to_fio2}()},
\code{\link{spo2_to_pao2}()}
}
\concept{helper}
