% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiration.r
\name{clean_fio2}
\alias{clean_fio2}
\title{Clean FiO2}
\usage{
clean_fio2(fio2, o2supply, o2system)
}
\arguments{
\item{fio2}{Fraction inspired oxygen (\%)}

\item{o2supply}{Oxygen supply (L/min)}

\item{o2system}{Oxygen supply method}
}
\value{
fio2 (\%)
}
\description{
Clean FiO2 values based on several rules.
}
\details{
Rules include cleaning on impossible values and on common mistakes
or switch-ups when entering data in the EHR.
}
\section{Todo}{

Implement more sophisticated FiO2 cleaning algorithm.
}

\seealso{
Other respiration: 
\code{\link{clean_spo2}()},
\code{\link{impute_fio2}()},
\code{\link{impute_pao2}()},
\code{\link{o2supply_to_fio2}()},
\code{\link{pfratio_imputed}()},
\code{\link{spo2_to_pao2}()}

Other cleaning: 
\code{\link{clean_spo2}()}
}
\concept{cleaning}
\concept{respiration}
