% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.r
\name{clean_lab}
\alias{clean_lab}
\title{Clean laboratory variables by removing '<' and '>' characters and converting to numeric
Cleans all columns starting with "lab_" (per Acutelines standard)}
\usage{
clean_lab(df)
}
\arguments{
\item{df}{Data frame containing laboratory variables}
}
\value{
Data frame with cleaned laboratory variables
}
\description{
Clean laboratory variables by removing '<' and '>' characters and converting to numeric
Cleans all columns starting with "lab_" (per Acutelines standard)
}
\seealso{
Other postprocessing: 
\code{\link{normalize_decimal_format}()}
}
\concept{postprocessing}
