% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiration.r
\name{clean_spo2}
\alias{clean_spo2}
\title{Clean SpO2}
\usage{
clean_spo2(spo2)
}
\arguments{
\item{spo2}{Peripheral oxygen saturation (\%)}
}
\value{
spo2 (\%)
}
\description{
Clean SpO2 values based on several rules.
}
\details{
Rules include cleaning on impossible values and on common mistakes
or switch-ups when entering data in the EHR.
Remove Spo2<50 and convert 50-70 to 70\%.
}
\seealso{
Other respiration: 
\code{\link{clean_fio2}()},
\code{\link{impute_fio2}()},
\code{\link{impute_pao2}()},
\code{\link{o2supply_to_fio2}()},
\code{\link{pfratio_imputed}()},
\code{\link{spo2_to_pao2}()}

Other cleaning: 
\code{\link{clean_fio2}()}
}
\concept{cleaning}
\concept{respiration}
