% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{impute_pao2}
\alias{impute_pao2}
\title{Impute PaO2 with SpO2. If PaO2 is missing replace with PaO2 calculated from SpO2}
\usage{
impute_pao2(pao2, spo2)
}
\arguments{
\item{pao2}{Arterial oxygen pressure (kPa)}

\item{spo2}{Peripheral oxygen saturation (\%)}
}
\value{
pao2
}
\description{
Impute PaO2 with SpO2. If PaO2 is missing replace with PaO2 calculated from SpO2
}
