% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.r
\name{news_consc}
\alias{news_consc}
\title{Calculate NEWS score for Consciousness level (AVPU or GCS)}
\usage{
news_consc(avpu = NA, gcs = NA)
}
\arguments{
\item{avpu}{Character value representing AVPU scale ("A", "V", "P", "U", "D")}

\item{gcs}{Numeric value representing Glasgow Coma Scale (3-15)}
}
\value{
Numeric NEWS score for Consciousness level
}
\description{
Prioritizes AVPU over GCS if both are provided.
}
\seealso{
Other news: 
\code{\link{news_hr}()},
\code{\link{news_o2supp}()},
\code{\link{news_rf}()},
\code{\link{news_sbp}()},
\code{\link{news_spo2}()},
\code{\link{news_temp}()},
\code{\link{news_total}()}
}
\concept{news}
