% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.r
\name{news_hr}
\alias{news_hr}
\title{Calculate NEWS score for Heart Rate (HR)}
\usage{
news_hr(hr)
}
\arguments{
\item{hr}{Numeric value representing Heart Rate (bpm)}
}
\value{
Numeric NEWS score for HR
}
\description{
Calculate NEWS score for Heart Rate (HR)
}
\seealso{
Other news: 
\code{\link{news_consc}()},
\code{\link{news_o2supp}()},
\code{\link{news_rf}()},
\code{\link{news_sbp}()},
\code{\link{news_spo2}()},
\code{\link{news_temp}()},
\code{\link{news_total}()}
}
\concept{news}
