% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.r
\name{news_o2supp}
\alias{news_o2supp}
\title{Calculate NEWS score for Oxygen Supplementation (O2 supp)}
\usage{
news_o2supp(o2supp)
}
\arguments{
\item{o2supp}{Logical value indicating if oxygen supplementation is provided (TRUE/FALSE)}
}
\value{
Numeric NEWS score for O2 supplementation
}
\description{
Calculate NEWS score for Oxygen Supplementation (O2 supp)
}
\seealso{
Other news: 
\code{\link{news_consc}()},
\code{\link{news_hr}()},
\code{\link{news_rf}()},
\code{\link{news_sbp}()},
\code{\link{news_spo2}()},
\code{\link{news_temp}()},
\code{\link{news_total}()}
}
\concept{news}
