% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.r
\name{news_rf}
\alias{news_rf}
\title{Calculate NEWS score for Respiratory Frequency (RF)}
\usage{
news_rf(rf)
}
\arguments{
\item{rf}{Numeric value representing Respiratory Frequency (breaths per minute)}
}
\value{
Numeric NEWS score for RF
}
\description{
Calculate NEWS score for Respiratory Frequency (RF)
}
\seealso{
Other news: 
\code{\link{news_consc}()},
\code{\link{news_hr}()},
\code{\link{news_o2supp}()},
\code{\link{news_sbp}()},
\code{\link{news_spo2}()},
\code{\link{news_temp}()},
\code{\link{news_total}()}
}
\concept{news}
