% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.r
\name{news_sbp}
\alias{news_sbp}
\title{Calculate NEWS score for Systolic Blood Pressure (SBP)}
\usage{
news_sbp(sbp)
}
\arguments{
\item{sbp}{Numeric value representing Systolic Blood Pressure (mmHg)}
}
\value{
Numeric NEWS score for SBP
}
\description{
Calculate NEWS score for Systolic Blood Pressure (SBP)
}
\seealso{
Other news: 
\code{\link{news_consc}()},
\code{\link{news_hr}()},
\code{\link{news_o2supp}()},
\code{\link{news_rf}()},
\code{\link{news_spo2}()},
\code{\link{news_temp}()},
\code{\link{news_total}()}
}
\concept{news}
