% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.r
\name{news_spo2}
\alias{news_spo2}
\title{Calculate NEWS score for Oxygen Saturation (SpO2)}
\usage{
news_spo2(spo2)
}
\arguments{
\item{spo2}{Numeric value representing Oxygen Saturation (\%)}
}
\value{
Numeric NEWS score for SpO2
}
\description{
Calculate NEWS score for Oxygen Saturation (SpO2)
}
\seealso{
Other news: 
\code{\link{news_consc}()},
\code{\link{news_hr}()},
\code{\link{news_o2supp}()},
\code{\link{news_rf}()},
\code{\link{news_sbp}()},
\code{\link{news_temp}()},
\code{\link{news_total}()}
}
\concept{news}
