% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.r
\name{news_temp}
\alias{news_temp}
\title{Calculate NEWS score for Temperature (Temp)}
\usage{
news_temp(temp)
}
\arguments{
\item{temp}{Numeric value representing Temperature (°C)}
}
\value{
Numeric NEWS score for Temp
}
\description{
Calculate NEWS score for Temperature (Temp)
}
\seealso{
Other news: 
\code{\link{news_consc}()},
\code{\link{news_hr}()},
\code{\link{news_o2supp}()},
\code{\link{news_rf}()},
\code{\link{news_sbp}()},
\code{\link{news_spo2}()},
\code{\link{news_total}()}
}
\concept{news}
