% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news.r
\name{news_total}
\alias{news_total}
\title{Calculate total NEWS score from individual components}
\usage{
news_total(hr, rf, temp, sbp, avpu, gcs, o2supp, spo2, return_df = FALSE)
}
\arguments{
\item{hr}{Numeric value representing Heart Rate (bpm)}

\item{rf}{Numeric value representing Respiratory Frequency (breaths per minute)}

\item{temp}{Numeric value representing Temperature (°C)}

\item{sbp}{Numeric value representing Systolic Blood Pressure (mmHg)}

\item{avpu}{Character value representing AVPU scale ("A", "V", "P", "U", "D")}

\item{gcs}{Numeric value representing Glasgow Coma Scale (3-15)}

\item{o2supp}{Logical value indicating if oxygen supplementation is provided (TRUE/FALSE)}

\item{spo2}{Numeric value representing Oxygen Saturation (\%)}

\item{return_df}{Logical value indicating if a detailed dataframe should be returned (TRUE/FALSE), default is FALSE}
}
\value{
Total NEWS score as an integer, or a dataframe with individual component scores and total score if return_df is TRUE
}
\description{
Calculate total NEWS score from individual components
}
\seealso{
Other news: 
\code{\link{news_consc}()},
\code{\link{news_hr}()},
\code{\link{news_o2supp}()},
\code{\link{news_rf}()},
\code{\link{news_sbp}()},
\code{\link{news_spo2}()},
\code{\link{news_temp}()}
}
\concept{news}
