% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.r
\name{normalize_decimal_format}
\alias{normalize_decimal_format}
\title{Normalize decimal format in a data frame
Interprets both "." and "," as decimal indicators}
\usage{
normalize_decimal_format(df)
}
\arguments{
\item{df}{Data frame to be normalized}
}
\value{
Data frame with normalized decimal format
}
\description{
Normalize decimal format in a data frame
Interprets both "." and "," as decimal indicators
}
\seealso{
Other postprocessing: 
\code{\link{clean_lab}()}
}
\concept{postprocessing}
