% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{o2supply_to_fio2}
\alias{o2supply_to_fio2}
\title{When O2 supply is below 16 consider it already FiO2}
\usage{
o2supply_to_fio2(o2supply)
}
\arguments{
\item{o2supply}{Oxygen supply}
}
\value{
fio2 fraction inspirated oxygen
}
\description{
When O2 supply is below 16 consider it already FiO2
}
