% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{pfratio_imputed}
\alias{pfratio_imputed}
\title{Calculate P/F-ratio}
\usage{
pfratio_imputed(pao2, spo2, fio2, oxygen_supply)
}
\arguments{
\item{pao2}{Arterial oxygen pressure (kPa)}

\item{spo2}{Peripheral oxygen saturation (\%)}

\item{fio2}{Fraction inspired oxygen}

\item{o2supply}{Oxygen supply (L/min)}
}
\value{
pfratio
}
\description{
Calculate P/F-ratio while imputing missing PaO2 using SpO2 and
missing FiO2 using oxygen supply (in L/min)
}
\details{
This function uses the impute_pao2 and impute_fio2 helper functions.
}
\seealso{
Other helper: 
\code{\link{cbrt}()},
\code{\link{impute_fio2}()},
\code{\link{impute_pao2}()},
\code{\link{o2supply_to_fio2}()},
\code{\link{spo2_to_pao2}()}
}
\concept{helper}
