% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_cardiovascular}
\alias{sofa_cardiovascular}
\title{Calculate the sofa score for cardiovacular
Note: due to inconsistent registration of norepinephrine dosages,
when norephenrine is present 3.5 points are given.
Researchers a free to either round this to 4 (overesitmate), 3 (underesitmate) or leave it as is}
\usage{
sofa_cardiovascular(
  map,
  dopamine,
  dobutamine,
  epinephrine,
  norepinephrine,
  norepinephrine_amp
)
}
\arguments{
\item{map}{Mean Arterial Pressure}

\item{dopamine}{Dopamine dosage}

\item{dobutamine}{Dobutamine dosage}

\item{epinephrine}{Epinephrine dosage}

\item{norepinephrine}{Norepinephrine dosage}

\item{norepinephrine_amp}{Norepinephrine ampul administration (1=TRUE, 0=FALSE)}
}
\value{
sofa score
}
\description{
Calculate the sofa score for cardiovacular
Note: due to inconsistent registration of norepinephrine dosages,
when norephenrine is present 3.5 points are given.
Researchers a free to either round this to 4 (overesitmate), 3 (underesitmate) or leave it as is
}
