% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_cns}
\alias{sofa_cns}
\title{Calculate the sofa score for Central Nervous System}
\usage{
sofa_cns(gcs)
}
\arguments{
\item{gcs}{Glascow Coma Scale}
}
\value{
sofa score
}
\description{
Calculate the sofa score for Central Nervous System
}
