% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_coagulation}
\alias{sofa_coagulation}
\title{Calculate sofa score for coagulation}
\usage{
sofa_coagulation(plt)
}
\arguments{
\item{plt}{Platelets (*10^3 uL^-1)}
}
\value{
sofa score
}
\description{
Calculate sofa score for coagulation
}
