% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_renal}
\alias{sofa_renal}
\title{Calculate the sofa score for renal
Note: Without urine output because it is unreliable}
\usage{
sofa_renal(creat)
}
\arguments{
\item{create}{Creatinine}
}
\value{
sofa score
}
\description{
Calculate the sofa score for renal
Note: Without urine output because it is unreliable
}
