% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_respiration}
\alias{sofa_respiration}
\title{Calculate sofa score for respiration
Needs either pfratio of both PaO2 and FiO2}
\usage{
sofa_respiration(pfratio, pao2, fio2)
}
\arguments{
\item{pfratio}{P/F-ratio (kPa)}

\item{pao2}{PaO2 (kPa)}

\item{fio2}{FiO2}
}
\value{
sofa score
}
\description{
Calculate sofa score for respiration
Needs either pfratio of both PaO2 and FiO2
}
