% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{spo2_to_pao2}
\alias{spo2_to_pao2}
\title{SOFA score calculation helper functions
Acutelines 2024}
\usage{
spo2_to_pao2(spo2)
}
\arguments{
\item{spo2}{SpO2 (\%) to be converted}
}
\value{
pao2 Calculated PaO2 (kPa)
}
\description{
Revisions:
v1 STH Initial script
v2 TvdA Revising calculations
v3 RvW Refactored code to functions, added Brown equation
Convert SpO2 to PaO2 based on Brown / Ellis equation
DOI: 10.1016/j.chest.2016.01.003
DOI: 0.1152/jappl.1979.46.3.599
}
