% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{impute_fio2}
\alias{impute_fio2}
\title{Impute FiO2}
\usage{
impute_fio2(fio2, o2supply)
}
\arguments{
\item{fio2}{Fraction inspired oxygen}

\item{o2supply}{Oxygen supply (L/min)}
}
\value{
fio2
}
\description{
Impute FiO2 with oxygen supply. If FiO2 is missing,
replace with FiO2 calculated from oxygen supply. If still
missing, impute with 21\% (ambient air).
}
\details{
FiO2 is missing in many cases. To fill those gaps, this function
converts the oxygen supply by nasal cannula (L/min) to FiO2
in case the FiO2 is missing in the dataset. When also the
oxygen supply is missing, the FiO2 is imputed with 21\%,
the ambient air oxygen percentage. This functions uses \code{\link[=o2supply_to_fio2]{o2supply_to_fio2()}}.
}
\seealso{
Other helper: 
\code{\link{cbrt}()},
\code{\link{impute_pao2}()},
\code{\link{o2supply_to_fio2}()},
\code{\link{pfratio_imputed}()},
\code{\link{spo2_to_pao2}()}
}
\concept{helper}
