% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{impute_fio2}
\alias{impute_fio2}
\title{Impute FiO2 with oxygen supply. If missing FiO2 replace with oxygen supply}
\usage{
impute_fio2(fio2, o2supply)
}
\arguments{
\item{fio2}{Fraction inspired oxygen}

\item{o2supply}{Oxygen supply (L/min)}
}
\value{
fio2
}
\description{
Impute FiO2 with oxygen supply. If missing FiO2 replace with oxygen supply
}
