% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/respiration.r
\name{pfratio_imputed}
\alias{pfratio_imputed}
\title{Calculate P/F-ratio}
\usage{
pfratio_imputed(pao2, spo2, fio2, o2supply, o2system)
}
\arguments{
\item{pao2}{Arterial oxygen pressure (kPa)}

\item{spo2}{Peripheral oxygen saturation (\%)}

\item{fio2}{Fraction inspired oxygen (\%)}

\item{o2supply}{Oxygen supply (L/min)}

\item{o2system}{Oxygen supply method (ranked 0-50)}
}
\value{
pfratio
}
\description{
Calculate P/F-ratio while cleaning and imputing PaO2 and FiO2.
}
\details{
This function uses \code{\link[=impute_pao2]{impute_pao2()}}, \code{\link[=clean_fio2]{clean_fio2()}} and \code{\link[=impute_fio2]{impute_fio2()}}.
}
\seealso{
Other respiration: 
\code{\link{clean_fio2}()},
\code{\link{clean_spo2}()},
\code{\link{impute_fio2}()},
\code{\link{impute_pao2}()},
\code{\link{o2supply_to_fio2}()},
\code{\link{spo2_to_pao2}()}
}
\concept{respiration}
