% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_cardiovascular}
\alias{sofa_cardiovascular}
\title{SOFA cardiovascular}
\usage{
sofa_cardiovascular(
  map,
  dopamine,
  dobutamine,
  epinephrine,
  norepinephrine,
  norepinephrine_amp
)
}
\arguments{
\item{map}{Mean Arterial Pressure}

\item{dopamine}{Dopamine dosage (ug/kg/min)}

\item{dobutamine}{Dobutamine dosage (ug/kg/min)}

\item{epinephrine}{Epinephrine dosage (ug/kg/min)}

\item{norepinephrine}{Norepinephrine dosage (ug/kg/min)}

\item{norepinephrine_amp}{Norepinephrine ampul
administration (1=TRUE, 0=FALSE)}
}
\value{
partial sofa score
}
\description{
This function calculates the cardiavascular part of the SOFA
score.
}
\details{
Due to inconsistent registration of norepinephrine dosages,
it's not possible to distinguish between high (>0.1) and
low (<=0.1) dosage of norepinephrine.
When norepinephrine is present 3.5 points are given.
Researchers can either round this to 4 (overestimate),
3 (underestimate) or leave it as is.
}
\seealso{
Other sofa: 
\code{\link{sofa_cns}()},
\code{\link{sofa_coagulation}()},
\code{\link{sofa_liver}()},
\code{\link{sofa_magic_wrapper}()},
\code{\link{sofa_renal}()},
\code{\link{sofa_respiration}()},
\code{\link{sofa_total}()}
}
\concept{sofa}
