% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_cns}
\alias{sofa_cns}
\title{SOFA CNS}
\usage{
sofa_cns(gcs)
}
\arguments{
\item{gcs}{Glascow Coma Scale}
}
\value{
partial sofa score
}
\description{
This function calculates the Central Nervous System (CNS) part of the SOFA
score.
}
\seealso{
Other sofa: 
\code{\link{sofa_cardiovascular}()},
\code{\link{sofa_coagulation}()},
\code{\link{sofa_liver}()},
\code{\link{sofa_magic_wrapper}()},
\code{\link{sofa_renal}()},
\code{\link{sofa_respiration}()},
\code{\link{sofa_total}()}
}
\concept{sofa}
