% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_coagulation}
\alias{sofa_coagulation}
\title{SOFA coagulation}
\usage{
sofa_coagulation(plt)
}
\arguments{
\item{plt}{Platelets (*10^3 uL^-1)}
}
\value{
partial sofa score
}
\description{
This function calculates the coagulation part of the SOFA
score. Platelet count need to be provided in *10^3 uL^-1.
}
\seealso{
Other sofa: 
\code{\link{sofa_cardiovascular}()},
\code{\link{sofa_cns}()},
\code{\link{sofa_liver}()},
\code{\link{sofa_magic_wrapper}()},
\code{\link{sofa_renal}()},
\code{\link{sofa_respiration}()},
\code{\link{sofa_total}()}
}
\concept{sofa}
