% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_liver}
\alias{sofa_liver}
\title{Calculate sofa score for liver}
\usage{
sofa_liver(bili)
}
\arguments{
\item{bili}{Bilirubin (umol L^-1)}
}
\value{
sofa score
}
\description{
Calculate sofa score for liver
}
