% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_renal}
\alias{sofa_renal}
\title{SOFA renal}
\usage{
sofa_renal(creat)
}
\arguments{
\item{create}{Creatinine (umol/L)}
}
\value{
partial sofa score
}
\description{
This function calculates the renal part of the SOFA
score.
}
\details{
This function ignores the urine output as defined in the SOFA score
Reliable measurements of urine output in a retrospective dataset of
both deteriorated and not deteriorated patients are scarce.
}
\seealso{
Other sofa: 
\code{\link{sofa_cardiovascular}()},
\code{\link{sofa_cns}()},
\code{\link{sofa_coagulation}()},
\code{\link{sofa_liver}()},
\code{\link{sofa_magic_wrapper}()},
\code{\link{sofa_respiration}()},
\code{\link{sofa_total}()}
}
\concept{sofa}
