% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofa.r
\name{sofa_respiration}
\alias{sofa_respiration}
\title{SOFA respiration}
\usage{
sofa_respiration(pfratio, pao2, fio2)
}
\arguments{
\item{pfratio}{P/F-ratio (kPa)}

\item{pao2}{PaO2 (kPa)}

\item{fio2}{FiO2}
}
\value{
partial sofa score
}
\description{
This function calculates the respiratory part of the SOFA
score. It needs either the PF-ratio or both PaO2 and FiO2.
If PF-ratio is missing, the funcion will calculate it from
PaO2 and FiO2.
}
\seealso{
Other sofa: 
\code{\link{sofa_cardiovascular}()},
\code{\link{sofa_cns}()},
\code{\link{sofa_coagulation}()},
\code{\link{sofa_liver}()},
\code{\link{sofa_magic_wrapper}()},
\code{\link{sofa_renal}()},
\code{\link{sofa_total}()}
}
\concept{sofa}
